BITを修正した

+++ data_structure/SegmentTree.h
@@ -13,10 +13,10 @@ private:
     const Op op;
     const T ID;
     std::vector<T> vec;
-    int sz;
+    size_t sz;
 
 public:
-    explicit SegmentTree(int _n, Op _op, T _identity) noexcept
+    explicit SegmentTree(size_t _n, Op _op, T _identity) noexcept
         : op(_op), ID(_identity) {
         sz = 1;
         while (sz < _n) sz <<= 1;
@@ -24,13 +24,13 @@ public:
     }
     explicit SegmentTree(const std::vector<T>& _vec, Op _op, T _identity) noexcept
         : op(_op), ID(_identity) {
-        int n = _vec.size();
+        size_t n = _vec.size();
         sz = 1;
         while (sz < n) sz <<= 1;
         vec.resize(sz * 2, ID);
         set_array(_vec);
     }
-    void set_value(int _idx, T _val) noexcept {
+    void set_value(size_t _idx, T _val) noexcept {
         vec[_idx + sz] = _val;
     }
     template <class RandomIt>
@@ -41,18 +41,19 @@ public:
         set_array(_newvec.begin(), _newvec.end());
     }
     void build() noexcept {
-        for (int i = sz - 1; i > 0; i--) {
+        for (size_t i = sz - 1; i > 0; i--) {
             vec[i] = op(vec[i * 2], vec[i * 2 + 1]);
         }
     }
-    void update(int _idx, T _val) noexcept {
+    void update(size_t _idx, T _val) noexcept {
         _idx += sz;
         vec[_idx] = _val;
         for (_idx >>= 1; _idx > 0; _idx >>= 1) {
             vec[_idx] = op(vec[_idx * 2], vec[_idx * 2 + 1]);
         }
     }
-    T query(int _l, int _r) const noexcept {
+    // 区間[_i, _r)に対するクエリを実行する
+    T query(size_t _l, size_t _r) const noexcept {
         T l_val = ID, r_val = ID;
         _l += sz, _r += sz - 1;
         for (; _l <= _r; _l >>= 1, _r >>= 1) {
@@ -61,7 +62,7 @@ public:
         }
         return op(l_val, r_val);
     }
-    const T& operator[](int _idx) const noexcept {
+    const T& operator[](size_t _idx) const noexcept {
         return vec[_idx + sz];
     }
 };
